requirements.txt
annotated-types==0.7.0
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
numpy==2.1.3
pandas==2.2.3
pdfminer.six==20240706
pycparser==2.22
pydantic==2.10.1
pydantic_core==2.27.1
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.16.0
typing_extensions==4.12.2
tzdata==2024.2
